requirements.txt
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
filelock==3.0.12
idna==2.10
joblib==1.0.0
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.1
Pillow==8.1.0
PySocks==1.7.1
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
schedule==0.6.0
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
tweepy==3.10.0
urllib3==1.26.2
virtualenv==20.4.2
discord~=1.0.1

